Fails horribly in yawed cases, even it there was initial yaw, had shown huzaifa bhai till here

+++ config.py
@@ -6,5 +6,5 @@ WIDTH = 10
 fx = fy = 800
 cx = cy = 400
 K = np.array([[fx, 0, cx], [0, fy, cy], [0, 0, 1]])
-distCoeffs = np.zeros((4,1))  # or use your actual distortion
+distCoeffs = np.zeros((4, 1), dtype=np.float32) # Don't change this line, this was after hours of debugging
 IMAGE_SIZE = (800, 800)
